Refactored the codebase by removing redundant and unused code to improve readability and maintainability. No functionality was affected
