--- src/main/java/com/store/book/dto/author/AuthorDto.java
@@ -1,10 +1,0 @@
-package com.store.book.dto.author;
-
-import lombok.Data;
-
-@Data
-public class AuthorDto {
-    private Long id;
-    private String name;
-    private String bio;
-}
