--- src/main/java/com/store/book/dto/author/CreateAuthorRequestDto.java
@@ -1,8 +1,0 @@
-package com.store.book.dto.author;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-public record CreateAuthorRequestDto(String name, String bio) {
-
-}
