+++ src/main/java/com/store/book/dto/book/BookDto.java
@@ -7,9 +7,9 @@ import lombok.Data;
 public class BookDto {
     private Long id;
     private String title;
+    private String author;
     private String isbn;
     private BigDecimal price;
     private String description;
     private String coverImage;
-    private Long authorId;
 }
