+++ src/main/java/com/store/book/dto/book/CreateBookRequestDto.java
@@ -1,6 +1,5 @@
 package com.store.book.dto.book;
 
-
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import java.math.BigDecimal;
 import lombok.Data;
@@ -9,9 +8,9 @@ import lombok.Data;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class CreateBookRequestDto {
     private String title;
+    private String author;
     private String isbn;
     private BigDecimal price;
     private String description;
     private String coverImage;
-    private Long authorId;
 }
