+++ src/main/java/com/store/book/mapper/BookMapper.java
@@ -5,15 +5,12 @@ import com.store.book.dto.book.BookDto;
 import com.store.book.dto.book.CreateBookRequestDto;
 import com.store.book.model.Book;
 import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
 
-@Mapper(config = MapperConfig.class, uses = AuthorMapper.class)
+@Mapper(config = MapperConfig.class)
 public interface BookMapper {
 
-    @Mapping(source = "author.id", target = "authorId")
     BookDto toDto(Book book);
 
-    @Mapping(target = "author", source = "authorId", qualifiedByName = "authorById")
     Book toModel(CreateBookRequestDto requestDto);
 
 }
