+++ src/main/java/com/store/book/model/Book.java
@@ -2,19 +2,15 @@ package com.store.book.model;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import lombok.ToString;
 import org.hibernate.annotations.Where;
 
 @Entity
@@ -33,6 +29,9 @@ public class Book {
     @Column(name = "title")
     private String title;
 
+    @Column(name = "author")
+    private String author;
+
     @Column(name = "isbn", unique = true)
     private String isbn;
 
@@ -45,12 +44,7 @@ public class Book {
     @Column(name = "coverImage")
     private String coverImage;
 
-    @EqualsAndHashCode.Exclude
-    @ToString.Exclude
-    @ManyToOne(fetch = FetchType.LAZY)
-    private Author author;
-
-    @Column(name = "deleted", nullable = false)
-    private boolean deleted = false;
+    @Column(name = "is_deleted", nullable = false)
+    private boolean isDeleted = false;
 
 }
