+++ src/main/java/com/store/book/service/BookService.java
@@ -1,0 +1,18 @@
+package com.store.book.service;
+
+import com.store.book.dto.book.BookDto;
+import com.store.book.dto.book.CreateBookRequestDto;
+import java.util.List;
+
+public interface BookService {
+    BookDto save(CreateBookRequestDto requestDto);
+
+    List<BookDto> findAll();
+
+    List<BookDto> findAllByTitle(String title);
+
+    BookDto findById(Long id);
+
+    void deleteById(Long id);
+
+}
