--- src/main/resources/db.changelog/changes/01-create-authors-table.yaml
@@ -1,23 +1,0 @@
-databaseChangeLog:
-  - changeSet:
-      id: create-authors-table
-      author: Volodymyr
-      changes:
-        - createTable:
-            tableName: authors
-            columns:
-              - column:
-                  name: id
-                  type: bigint
-                  autoIncrement: true
-                  constraints:
-                    primaryKey: true
-                    nullable: false
-              - column:
-                  name: name
-                  type: varchar(255)
-                  constraints:
-                    nullable: false
-              - column:
-                  name: bio
-                  type: varchar(255)
