--- src/main/resources/db.changelog/changes/02-create-books-table.yaml
@@ -1,49 +1,0 @@
-databaseChangeLog:
-  - changeSet:
-      id: create-book-table
-      author: Volodymyr
-      changes:
-        - createTable:
-            tableName: books
-            columns:
-              - column:
-                  name: id
-                  type: bigint
-                  autoIncrement: true
-                  constraints:
-                    primaryKey: true
-                    nullable: false
-              - column:
-                  name: title
-                  type: varchar(255)
-              - column:
-                  name: isbn
-                  type: varchar(255)
-                  constraints:
-                    unique: true
-              - column:
-                  name: price
-                  type: decimal(19, 2)
-              - column:
-                  name: description
-                  type: varchar(255)
-              - column:
-                  name: cover_image
-                  type: varchar(255)
-              - column:
-                  name: is_deleted
-                  type: boolean
-                  defaultValueBoolean: false
-                  constraints:
-                    nullable: false
-              - column:
-                  name: author_id
-                  type: bigint
-                  constraints:
-                    nullable: false
-        - addForeignKeyConstraint:
-            baseTableName: books
-            baseColumnNames: author_id
-            referencedTableName: authors
-            referencedColumnNames: id
-            constraintName: fk_books_authors
