+++ src/main/resources/db.changelog/db.changelog-master.yaml
@@ -1,5 +1,3 @@
 databaseChangeLog:
   - include:
-      file: db/changelog/changes/01-create-authors-table.yaml
-  - include:
-      file: db/changelog/changes/02-create-books-table.yamlю
+      file: db/changelog/changes/01-create-books-table.yaml
